+++ public/index.html
@@ -8,7 +8,7 @@
     <title>Ro1ling</title>
 
     <!-- 미리보기 -->
-    <meta property="og:image" content="%PUBLIC_URL%/preview.png" />
+    <meta property="og:image" content="%PUBLIC_URL%/preview.webp" />
     <meta property="og:title" content="Ro1ling" />
     <meta
       property="og:description"
@@ -18,7 +18,7 @@
 
     <!-- 트위터 미리보기 -->
     <meta name="twitter:card" content="summary_large_image" />
-    <meta name="twitter:image" content="%PUBLIC_URL%/preview.png" />
+    <meta name="twitter:image" content="%PUBLIC_URL%/preview.webp" />
     <meta property="twitter:title" content="Ro1ling" />
     <meta
       property="twitter:description"
