+++ src/components/ListPage/CardList.module.scss
@@ -41,19 +41,19 @@
     }
 
     &.bg-purple {
-      background-image: url("../../assets/images/purple.png");
+      background-image: url("../../assets/images/purple.webp");
     }
     
     &.bg-beige {
-      background-image: url("../../assets/images/beige.png");
+      background-image: url("../../assets/images/beige.webp");
     }
     
     &.bg-blue {
-      background-image: url("../../assets/images/blue.png");
+      background-image: url("../../assets/images/blue.webp");
     }
     
     &.bg-green {
-      background-image: url("../../assets/images/green.png");
+      background-image: url("../../assets/images/green.webp");
     }
   }
